Accumulate PGP dumb component maps into subfolder
